+++ service/src/main/kotlin/fi/espoo/evaka/vasu/VasuController.kt
@@ -4,17 +4,14 @@ import fi.espoo.evaka.shared.auth.AuthenticatedUser
 import fi.espoo.evaka.shared.auth.UserRole
 import fi.espoo.evaka.shared.db.Database
 import fi.espoo.evaka.shared.domain.BadRequest
-import fi.espoo.evaka.shared.domain.FiniteDateRange
 import fi.espoo.evaka.shared.domain.HelsinkiDateTime
 import fi.espoo.evaka.shared.domain.NotFound
-import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PathVariable
 import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.PutMapping
 import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RestController
-import java.net.URI
 import java.time.LocalDate
 import java.time.LocalDateTime
 import java.util.UUID
@@ -31,10 +28,10 @@ class VasuController {
         db: Database.Connection,
         user: AuthenticatedUser,
         @RequestBody body: CreateDocumentRequest
-    ): ResponseEntity<VasuDocumentResponse> {
+    ): VasuDocumentResponse {
         user.requireOneOfRoles(UserRole.ADMIN)
 
-        val document = db.transaction { tx ->
+        return db.transaction { tx ->
             tx.getVasuTemplate(body.templateId)?.let { template ->
                 if (!template.valid.includes(LocalDate.now())) {
                     throw BadRequest("Template is not currently valid")
@@ -44,8 +41,6 @@ class VasuController {
             val id = tx.insertVasuDocument(body.childId, body.templateId)
             tx.getVasuDocumentResponse(id)!!
         }
-
-        return ResponseEntity.created(URI.create("/vasu/${document.id}")).body(document)
     }
 
     @GetMapping("/vasu/{id}")
@@ -74,60 +69,6 @@ class VasuController {
         user.requireOneOfRoles(UserRole.ADMIN)
 
         db.transaction { tx -> tx.updateVasuDocument(id, body.content) }
-    }
-
-    data class CreateTemplateRequest(
-        val name: String,
-        val valid: FiniteDateRange
-    )
-    @PostMapping("/vasu/templates")
-    fun postTemplate(
-        db: Database.Connection,
-        user: AuthenticatedUser,
-        @RequestBody body: CreateTemplateRequest
-    ): ResponseEntity<UUID> {
-        user.requireOneOfRoles(UserRole.ADMIN)
-
-        return db.transaction {
-            it.insertVasuTemplate(
-                name = body.name,
-                valid = body.valid,
-                content = defaultTemplateContent
-            )
-        }.let { ResponseEntity.created(URI.create("/vasu/templates/$it")).body(it) }
-    }
-
-    @GetMapping("/vasu/templates")
-    fun getTemplates(
-        db: Database.Connection,
-        user: AuthenticatedUser
-    ): List<VasuTemplateSummary> {
-        user.requireOneOfRoles(UserRole.ADMIN)
-
-        return db.read { tx -> tx.getVasuTemplates() }
-    }
-
-    @GetMapping("/vasu/templates/{id}")
-    fun getTemplate(
-        db: Database.Connection,
-        user: AuthenticatedUser,
-        @PathVariable id: UUID
-    ): VasuTemplate {
-        user.requireOneOfRoles(UserRole.ADMIN)
-
-        return db.read { tx -> tx.getVasuTemplate(id) } ?: throw NotFound("template $id not found")
-    }
-
-    @PutMapping("/vasu/templates/{id}/content")
-    fun putTemplateContent(
-        db: Database.Connection,
-        user: AuthenticatedUser,
-        @PathVariable id: UUID,
-        @RequestBody content: VasuContent
-    ) {
-        user.requireOneOfRoles(UserRole.ADMIN)
-
-        db.transaction { tx -> tx.updateVasuTemplateContent(id, content) }
     }
 
     @GetMapping("/children/{childId}/vasu-summaries")
